sdk/trace: use sync.Pool in randomIDGenerator instead of Mutex
randomIDGenerator used a sync.Mutex to coordinate generation of TraceID
and SpanID. When used in relatively hot code paths this would cause
significant mutex contention as observed in pprof/mutex profiles.

This uses a sync.Pool instead which allows using generating and using
of *rand.Rand as needed without the need for a Lock.

Signed-off-by: Maisem Ali <maisem@tailscale.com>